+++ Emuera/Runtime/Script/Statements/Instraction.Child.cs
@@ -1318,6 +1318,12 @@ internal sealed partial class FunctionIdentifier
 			SpCallSharpArgment arg = (SpCallSharpArgment)func.Argument;
 			var manager = PluginManager.GetInstance();
 
+			if (arg.CallFunc == null)
+			{
+				exm.VEvaluator.RESULT = 0;
+				return;
+			}
+
 			var pluginArgs = arg.RowArgs.Select((term) => PluginMethodParameterBuilder.ConvertTerm(term, exm)).ToArray();
 			arg.CallFunc.Execute(pluginArgs);
 			for (var i = 0; i < pluginArgs.Count(); ++i)
@@ -1460,7 +1466,7 @@ internal sealed partial class FunctionIdentifier
 				catch (OverflowException)
 				{
 					GlobalStatic.EMediator.Console.PrintWarning(
-						$"TIMES整数溢出: {d}", null, 1);
+						$"TIMES整数溢出: {d}", default(ScriptPosition), 1);
 					var.SetValue(d > 0 ? long.MaxValue : long.MinValue, exm);
 				}
 			}
@@ -1472,7 +1478,7 @@ internal sealed partial class FunctionIdentifier
 				else
 				{
 					GlobalStatic.EMediator.Console.PrintWarning(
-						$"TIMES整数溢出: {d}", null, 1);
+						$"TIMES整数溢出: {d}", default(ScriptPosition), 1);
 					var.SetValue(d > 0 ? long.MaxValue : long.MinValue, exm);
 				}
 			}